move its to tests folder
