--- its/.gitignore
@@ -1,1 +1,0 @@
-.its-venv/
