--- its/pyproject.toml
@@ -1,60 +1,0 @@
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
-
-[project]
-# PYSCAN-35: Decide on definitive package name
-name = "pysonar-scanner-its"
-version = "0.0.1"
-authors = [{ name="[Code Quality] Data & ML Squad", email="quality.data-ml@sonarsource.com" }]
-description = "Integration tests for Sonar Scanner for the Python Ecosystem"
-readme = "README.md"
-license = "LGPL-3.0-only"
-keywords = [
-  "sonar",
-  "sonarqube",
-  "sonarcloud",
-  "cleancode"
-]
-classifiers = [
-  "Environment :: Console",
-  "Intended Audience :: Developers",
-  "Operating System :: OS Independent",
-  "Programming Language :: Python",
-  "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3 :: Only",
-  "Programming Language :: Python :: 3.9",
-  "Programming Language :: Python :: 3.10",
-  "Programming Language :: Python :: 3.11",
-  "Programming Language :: Python :: 3.12",
-  "Programming Language :: Python :: 3.13",
-  "Topic :: Software Development :: Quality Assurance",
-]
-
-dynamic = [ "dependencies" ]
-
-[tool.poetry.dependencies]
-python = ">=3.9"
-licenseheaders = ">=0.8.8"
-black = "25.1.0"
-pytest = ">=7.4.3"
-pytest-docker = ">=2.0.1"
-requests = "^2.32.3"
-pysonar-scanner = { "path" = "../" }
-
-[[tool.poetry.packages]]
-include = "utils"
-
-[[tool.poetry.source]]
-name = 'jfrog-server'
-url = 'https://repox.jfrog.io/artifactory/api/pypi/sonarsource-pypi/simple'
-priority = "primary"
-
-[tool.pytest.ini_options]
-addopts = [
-  "--import-mode=importlib",
-]
-
-[tool.black]
-line-length = 120
-target-version = ["py38", "py39", "py310", "py311", "py312"]
