--- its/sources/minimal/sonar-project.properties
@@ -1,1 +1,0 @@
-sonar.projectVersion=1.2
