--- its/sources/minimal/src/minimal.py
@@ -1,2 +1,0 @@
-def foo(test):
-    print("42")
