+++ pyproject.toml
@@ -45,6 +45,7 @@ mypy = '>=1.7.1'
 pytest = '8.3.5'
 pytest-cov = '6.0.0'
 pytest-subtests = "^0.14.1"
+pytest-docker = "^3.2.0"
 
 [[tool.poetry.packages]]
 from = 'src'
@@ -60,5 +61,8 @@ priority = "primary"
 
 [tool.pytest]
 [tool.pytest.ini_options]
-addopts = ['--import-mode=importlib']
+addopts = ['--import-mode=importlib', '--strict-markers']
 pythonpath = ['src']
+markers = [
+    "its: marks tests as its"
+]
