+++ tests/its/sources/minimal/pyproject.toml
