+++ tests/its/sources/minimal/sonar-project.properties
@@ -1,0 +1,1 @@
+sonar.projectVersion=1.2
