+++ tests/its/test_minimal.py
@@ -17,8 +17,13 @@
 # along with this program; if not, write to the Free Software Foundation,
 # Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #
-from utils.sonarqube_client import SonarQubeClient
-from utils.cli_client import CliClient
+import pytest
+from tests.its.utils.sonarqube_client import SonarQubeClient
+from tests.its.utils.cli_client import CliClient
+
+
+# Marks this module
+pytestmark = pytest.mark.its
 
 
 def test_minimal_project(sonarqube_client: SonarQubeClient, cli: CliClient):
