+++ tests/unit/test_sonar_project_properties.py
